Fix for Quills issue #193: post viewed by archive URL are rendered without CSS styling in NuPlone. This is a version of issue #97, this time for QuillsEnabled.
This was caused by NuPlone using IContextState.is_view_template to
detect 'view'-views. The default implementation will not work with
virtual URLs. Products.Quills already solves this problem by providing
a customized 'plone_context_state' view (r50053). This view is now
used by QuillsEnabled also. To prevent this view from accidentally
triggering for Plone documents outside weblogs, it is registered for
the IInsideWeblog layer. The implementation of is_view_template
requires a defaultView name to be set. This done also relative to the
same layer.

The doctest merely contains a description of the problem. I am unable
to devise a test, yet.